+++ api/genesis.go
@@ -1,0 +1,9 @@
+package api
+
+import (
+	"github.com/tendermint/tendermint/rpc/core/types"
+)
+
+func Genesis() (*core_types.ResultGenesis, error) {
+	return client.Genesis()
+}