+++ core/code/code.go
@@ -45,6 +45,7 @@ const (
 	CheckUsed        uint32 = 503
 	TooHighGasPrice  uint32 = 504
 	WrongGasCoin     uint32 = 505
+	TooLongNonce     uint32 = 506
 
 	// multisig
 	IncorrectWeights        uint32 = 601