+++ core/transaction/buy_coin.go
@@ -305,8 +305,8 @@ func (data BuyCoinData) Run(tx *Transaction, context *state.StateDB, isCheck boo
 		context.AddBalance(sender, data.CoinToBuy, data.ValueToBuy)
 		context.SetNonce(sender, tx.Nonce)
 
-		context.DeleteCoinIfZeroReserve(data.CoinToBuy)
-		context.DeleteCoinIfZeroReserve(data.CoinToSell)
+		context.SanitizeCoin(data.CoinToBuy)
+		context.SanitizeCoin(data.CoinToSell)
 	}
 
 	tags := common.KVPairs{