+++ core/transaction/executor.go
@@ -150,7 +150,7 @@ func RunTx(context *state.StateDB, isCheck bool, rawTx []byte, rewardPool *big.I
 	response.GasPrice = tx.GasPrice
 
 	if !isCheck && response.Code == code.OK {
-		context.DeleteCoinIfZeroReserve(tx.GasCoin)
+		context.SanitizeCoin(tx.GasCoin)
 	}
 
 	return response