+++ genesis/genesis.go
@@ -14,7 +14,7 @@ import (
 )
 
 var (
-	Network     = "minter-test-network-34"
+	Network     = "minter-test-network-35"
 	genesisTime = time.Date(2019, time.March, 27, 12, 0, 0, 0, time.UTC)
 
 	BlockMaxBytes int64 = 10000000