+++ scripts/dist.sh
@@ -23,11 +23,6 @@ mkdir -p build/pkg
 GIT_COMMIT="$(git rev-parse --short=8 HEAD)"
 GIT_IMPORT="github.com/MinterTeam/minter-go-node/version"
 
-# Determine the arch/os combos we're building for
-XC_ARCH=${XC_ARCH:-"amd64"}
-XC_OS=${XC_OS:-"darwin linux windows"}
-XC_EXCLUDE=${XC_EXCLUDE:-" darwin/arm solaris/amd64 solaris/386 solaris/arm freebsd/amd64 windows/arm "}
-
 # Make sure build tools are available.
 make get_tools
 
@@ -39,17 +34,11 @@ make get_vendor_deps
 # Build!
 # ldflags: -s Omit the symbol table and debug information.
 #	         -w Omit the DWARF symbol table.
-echo "==> Building..."
-IFS=' ' read -ra arch_list <<< "$XC_ARCH"
-IFS=' ' read -ra os_list <<< "$XC_OS"
-for arch in "${arch_list[@]}"; do
-	for os in "${os_list[@]}"; do
-	    if [[ "$XC_EXCLUDE" !=  *" $os/$arch "* ]]; then
-            echo "--> $os/$arch"
-            GOOS=${os} GOARCH=${arch} go build -ldflags "-s -w -X ${GIT_IMPORT}.GitCommit=${GIT_COMMIT}" -tags="${BUILD_TAGS}" -o "build/pkg/${os}_${arch}/minter" ./cmd/minter
-	    fi
-	done
-done
+echo "==> Building for mac os..."
+CGO_ENABLED=1 CGO_LDFLAGS="-lsnappy" go build -tags "minter gcc" -ldflags "-s -w -X ${GIT_IMPORT}.GitCommit=${GIT_COMMIT}" -o "build/pkg/darwin_amd64/minter" ./cmd/minter
+
+echo "==> Building for linux in docker"
+docker run -t -v ${PWD}:/go/src/github.com/MinterTeam/minter-go-node/ -i 37effcb3746d sh -c 'CGO_ENABLED=1 CGO_LDFLAGS="-lsnappy" go build -tags "minter gcc" -ldflags "-s -w -X ${GIT_IMPORT}.GitCommit=${GIT_COMMIT}" -o "build/pkg/linux_amd64/minter" ./cmd/minter/'
 
 # Zip all the files.
 echo "==> Packaging..."